[CHORE] #89 : Management 모듈에서 Management-event 모듈 분리
