+++ .idea/gradle.xml
@@ -16,7 +16,6 @@
             </builds>
           </compositeBuild>
         </compositeConfiguration>
-        <option name="testRunner" value="GRADLE" />
         <option name="externalProjectPath" value="$PROJECT_DIR$" />
         <option name="gradleJvm" value="jbr-17" />
         <option name="modules">
@@ -36,12 +35,14 @@
             <option value="$PROJECT_DIR$/feature" />
             <option value="$PROJECT_DIR$/feature/auth" />
             <option value="$PROJECT_DIR$/feature/management" />
+            <option value="$PROJECT_DIR$/feature/management-event" />
             <option value="$PROJECT_DIR$/feature/management-survey" />
             <option value="$PROJECT_DIR$/feature/notice" />
             <option value="$PROJECT_DIR$/feature/profile" />
             <option value="$PROJECT_DIR$/feature/splash" />
             <option value="$PROJECT_DIR$/feature/survey" />
             <option value="$PROJECT_DIR$/feature/survey-check" />
+            <option value="$PROJECT_DIR$/management-event" />
           </set>
         </option>
         <option name="resolveExternalAnnotations" value="false" />
