+++ app/src/main/java/com/wap/wapp/MainActivity.kt
@@ -25,7 +25,7 @@ import com.wap.wapp.component.WappBottomBar
 import com.wap.wapp.core.domain.usecase.auth.SignInUseCase
 import com.wap.wapp.feature.auth.signin.navigation.signInNavigationRoute
 import com.wap.wapp.feature.auth.signup.navigation.signUpNavigationRoute
-import com.wap.wapp.feature.management.registration.event.navigation.eventRegistrationNavigationRoute
+import com.wap.wapp.feature.management.event.navigation.eventRegistrationNavigationRoute
 import com.wap.wapp.feature.management.survey.registration.navigation.surveyRegistrationNavigationRoute
 import com.wap.wapp.feature.profile.profilesetting.navigation.profileSettingNavigationRoute
 import com.wap.wapp.feature.splash.navigation.splashNavigationRoute
