+++ app/src/main/java/com/wap/wapp/navigation/WappNavHost.kt
@@ -12,12 +12,11 @@ import com.wap.wapp.feature.auth.signin.navigation.navigateToSignIn
 import com.wap.wapp.feature.auth.signin.navigation.signInScreen
 import com.wap.wapp.feature.auth.signup.navigation.navigateToSignUp
 import com.wap.wapp.feature.auth.signup.navigation.signUpScreen
-import com.wap.wapp.feature.management.edit.event.navigation.eventEditScreen
-import com.wap.wapp.feature.management.edit.event.navigation.navigateToEventEdit
+import com.wap.wapp.feature.management.event.navigation.managementEventNavGraph
+import com.wap.wapp.feature.management.event.navigation.navigateToEventEdit
+import com.wap.wapp.feature.management.event.navigation.navigateToEventRegistration
 import com.wap.wapp.feature.management.navigation.managementScreen
 import com.wap.wapp.feature.management.navigation.navigateToManagement
-import com.wap.wapp.feature.management.registration.event.navigation.eventRegistrationScreen
-import com.wap.wapp.feature.management.registration.event.navigation.navigateToEventRegistration
 import com.wap.wapp.feature.management.survey.registration.navigation.navigateToSurveyRegistration
 import com.wap.wapp.feature.management.survey.registration.navigation.surveyRegistrationScreen
 import com.wap.wapp.feature.notice.navigation.navigateToNotice
@@ -78,10 +77,7 @@ fun WappNavHost(
         surveyRegistrationScreen(
             navigateToManagement = navController::navigateToManagement,
         )
-        eventRegistrationScreen(
-            navigateToManagement = navController::navigateToManagement,
-        )
-        eventEditScreen(
+        managementEventNavGraph(
             navigateToManagement = navController::navigateToManagement,
         )
         profileScreen(
