+++ feature/management-event/proguard-rules.pro
