+++ feature/management-event/src/main/AndroidManifest.xml
