+++ feature/management-event/src/main/java/com/wap/wapp/feature/management/event/component/RegistrationTitle.kt
@@ -1,4 +1,4 @@
-package com.wap.wapp.feature.management.registration.component
+package com.wap.wapp.feature.management.event.component
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
