+++ feature/management-event/src/main/java/com/wap/wapp/feature/management/event/component/WappDatePickerDialog.kt
@@ -1,4 +1,4 @@
-package com.wap.wapp.feature.management.registration.component
+package com.wap.wapp.feature.management.event.component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -31,7 +31,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
 import com.wap.designsystem.WappTheme
 import com.wap.wapp.core.commmon.util.DateUtil
-import com.wap.wapp.feature.management.R
+import com.wap.wapp.feature.management.event.R
 import java.time.DayOfWeek
 import java.time.LocalDate
 import java.time.format.TextStyle
