+++ feature/management-event/src/main/java/com/wap/wapp/feature/management/event/edit/EventEditScreen.kt
@@ -1,4 +1,4 @@
-package com.wap.wapp.feature.management.edit.event
+package com.wap.wapp.feature.management.event.edit
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -31,10 +31,9 @@ import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.wap.designsystem.WappTheme
 import com.wap.designsystem.component.WappSubTopBar
 import com.wap.wapp.core.commmon.extensions.toSupportingText
-import com.wap.wapp.feature.management.R
-import com.wap.wapp.feature.management.registration.event.EventRegistrationContent
-import com.wap.wapp.feature.management.registration.event.EventRegistrationEvent
-import com.wap.wapp.feature.management.registration.event.EventRegistrationState
+import com.wap.wapp.feature.management.event.R
+import com.wap.wapp.feature.management.event.registration.EventRegistrationEvent
+import com.wap.wapp.feature.management.event.registration.EventRegistrationState
 import kotlinx.coroutines.flow.collectLatest
 import java.time.LocalDate
 import java.time.LocalTime
@@ -162,7 +161,7 @@ internal fun EventEditScreen(
                 modifier = Modifier.padding(top = 16.dp),
             )
 
-            EventRegistrationContent(
+            com.wap.wapp.feature.management.event.registration.EventRegistrationContent(
                 eventRegistrationState = currentEditState,
                 modifier = Modifier.padding(top = 50.dp),
                 eventTitle = title,
