+++ feature/management-event/src/main/java/com/wap/wapp/feature/management/event/navigation/ManagementEventNavigation.kt
@@ -1,4 +1,4 @@
-package com.wap.wapp.feature.management.edit.event.navigation
+package com.wap.wapp.feature.management.event.navigation
 
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
@@ -7,9 +7,15 @@ import androidx.navigation.NavType
 import androidx.navigation.compose.composable
 import androidx.navigation.navArgument
 import androidx.navigation.navOptions
-import com.wap.wapp.feature.management.edit.event.EventEditRoute
+import com.wap.wapp.feature.management.event.edit.EventEditRoute
+import com.wap.wapp.feature.management.event.registration.EventRegistrationRoute
 
+const val eventRegistrationNavigationRoute = "event_registration_route"
 const val eventEditNavigationRoute = "event/edit/{date}/{eventId}"
+
+fun NavController.navigateToEventRegistration(navOptions: NavOptions? = navOptions {}) {
+    this.navigate(eventRegistrationNavigationRoute, navOptions)
+}
 
 fun NavController.navigateToEventEdit(
     date: String,
@@ -19,9 +25,15 @@ fun NavController.navigateToEventEdit(
     this.navigate("event/edit/$date/$eventId", navOptions)
 }
 
-fun NavGraphBuilder.eventEditScreen(
+fun NavGraphBuilder.managementEventNavGraph(
     navigateToManagement: () -> Unit,
 ) {
+    composable(route = eventRegistrationNavigationRoute) {
+        EventRegistrationRoute(
+            navigateToManagement = navigateToManagement,
+        )
+    }
+
     composable(
         route = eventEditNavigationRoute,
         arguments = listOf(
