+++ feature/management-event/src/main/java/com/wap/wapp/feature/management/event/registration/EventRegistrationViewModel.kt
@@ -1,12 +1,12 @@
-package com.wap.wapp.feature.management.edit.event
+package com.wap.wapp.feature.management.event.registration
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.wap.wapp.core.commmon.util.DateUtil
-import com.wap.wapp.core.domain.usecase.event.GetEventUseCase
-import com.wap.wapp.core.domain.usecase.event.UpdateEventUseCase
-import com.wap.wapp.feature.management.registration.event.EventRegistrationEvent
-import com.wap.wapp.feature.management.registration.event.EventRegistrationState
+import com.wap.wapp.core.commmon.util.DateUtil.generateNowDate
+import com.wap.wapp.core.commmon.util.DateUtil.generateNowTime
+import com.wap.wapp.core.domain.usecase.event.RegisterEventUseCase
+import com.wap.wapp.feature.management.event.registration.EventRegistrationState.EVENT_DETAILS
+import com.wap.wapp.feature.management.event.registration.EventRegistrationState.EVENT_SCHEDULE
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -14,23 +14,20 @@ import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
 import java.time.LocalDate
-import java.time.LocalDateTime
 import java.time.LocalTime
-import java.time.format.DateTimeFormatter
 import javax.inject.Inject
 
 @HiltViewModel
-class EventEditViewModel @Inject constructor(
-    private val getEventUseCase: GetEventUseCase,
-    private val updateEventUseCase: UpdateEventUseCase,
+class EventRegistrationViewModel @Inject constructor(
+    private val registerEventUseCase: RegisterEventUseCase,
 ) : ViewModel() {
-    private val _currentEditState: MutableStateFlow<EventRegistrationState> =
-        MutableStateFlow(EventRegistrationState.EVENT_DETAILS)
-    val currentEditState = _currentEditState.asStateFlow()
+    private val _currentRegistrationState: MutableStateFlow<EventRegistrationState> =
+        MutableStateFlow(EVENT_DETAILS)
+    val currentRegistrationState = _currentRegistrationState.asStateFlow()
 
-    private val _eventEditEvent: MutableSharedFlow<EventRegistrationEvent> =
+    private val _eventRegistrationEvent: MutableSharedFlow<EventRegistrationEvent> =
         MutableSharedFlow()
-    val eventEditEvent = _eventEditEvent.asSharedFlow()
+    val eventRegistrationEvent = _eventRegistrationEvent.asSharedFlow()
 
     private val _eventTitle: MutableStateFlow<String> = MutableStateFlow("")
     val eventTitle = _eventTitle.asStateFlow()
@@ -42,22 +39,20 @@ class EventEditViewModel @Inject constructor(
     val eventLocation = _eventLocation.asStateFlow()
 
     private val _eventStartDate: MutableStateFlow<LocalDate> =
-        MutableStateFlow(DateUtil.generateNowDate())
+        MutableStateFlow(generateNowDate())
     val eventStartDate = _eventStartDate.asStateFlow()
 
     private val _eventStartTime: MutableStateFlow<LocalTime> =
-        MutableStateFlow(DateUtil.generateNowTime())
+        MutableStateFlow(generateNowTime())
     val eventStartTime = _eventStartTime.asStateFlow()
 
     private val _eventEndDate: MutableStateFlow<LocalDate> =
-        MutableStateFlow(DateUtil.generateNowDate())
+        MutableStateFlow(generateNowDate())
     val eventEndDate = _eventEndDate.asStateFlow()
 
     private val _eventEndTime: MutableStateFlow<LocalTime> =
-        MutableStateFlow(DateUtil.generateNowTime().plusHours(1))
+        MutableStateFlow(generateNowTime().plusHours(1))
     val eventEndTime = _eventEndTime.asStateFlow()
-
-    private val _eventId: MutableStateFlow<String> = MutableStateFlow("")
 
     fun setEventTitle(eventTitle: String) {
         _eventTitle.value = eventTitle
@@ -72,6 +67,10 @@ class EventEditViewModel @Inject constructor(
     }
 
     fun setEventStartDate(eventDate: LocalDate) {
+        if (!isValidStartDate(eventDate)) {
+            emitValidationErrorMessage("최소 하루 이상 일정 날짜를 지정하세요.")
+            return
+        }
         _eventStartDate.value = eventDate
     }
 
@@ -96,7 +95,7 @@ class EventEditViewModel @Inject constructor(
     }
 
     fun setEventRegistrationState() {
-        if (_currentEditState.value == EventRegistrationState.EVENT_DETAILS) {
+        if (_currentRegistrationState.value == EVENT_DETAILS) {
             if (!isValidTitle()) {
                 emitValidationErrorMessage("행사 이름을 입력하세요.")
                 return
@@ -105,18 +104,17 @@ class EventEditViewModel @Inject constructor(
                 emitValidationErrorMessage("행사 내용을 입력하세요.")
                 return
             }
-            _currentEditState.value = EventRegistrationState.EVENT_SCHEDULE
+            _currentRegistrationState.value = EVENT_SCHEDULE
         }
     }
 
-    fun updateEvent() {
+    fun registerEvent() {
         if (!isValidLocation()) {
             emitValidationErrorMessage("장소를 입력하세요.")
             return
         }
-
         viewModelScope.launch {
-            updateEventUseCase(
+            registerEventUseCase(
                 eventTitle = _eventTitle.value,
                 eventContent = _eventContent.value,
                 eventLocation = _eventLocation.value,
@@ -124,11 +122,10 @@ class EventEditViewModel @Inject constructor(
                 eventStartTime = _eventStartTime.value,
                 eventEndDate = _eventEndDate.value,
                 eventEndTime = _eventEndTime.value,
-                eventId = _eventId.value,
             ).onSuccess {
-                _eventEditEvent.emit(EventRegistrationEvent.Success)
+                _eventRegistrationEvent.emit(EventRegistrationEvent.Success)
             }.onFailure { throwable ->
-                _eventEditEvent.emit(EventRegistrationEvent.Failure(throwable))
+                _eventRegistrationEvent.emit(EventRegistrationEvent.Failure(throwable))
             }
         }
     }
@@ -138,30 +135,17 @@ class EventEditViewModel @Inject constructor(
 
     private fun isValidEndDate(eventDate: LocalDate): Boolean = eventDate >= _eventStartDate.value
 
+    private fun isValidStartDate(eventDate: LocalDate): Boolean = eventDate > generateNowDate()
+
     private fun isValidContent(): Boolean = _eventContent.value.isNotEmpty()
 
     private fun isValidTitle(): Boolean = _eventTitle.value.isNotEmpty()
 
     private fun isValidLocation(): Boolean = _eventLocation.value.isNotEmpty()
 
-    fun getEvent(date: String, eventId: String) = viewModelScope.launch {
-        val date = LocalDateTime.parse(date, DateTimeFormatter.ISO_LOCAL_DATE_TIME)
-        getEventUseCase(date, eventId).onSuccess {
-            _eventContent.value = it.content
-            _eventTitle.value = it.title
-            _eventStartDate.value = it.startDateTime.toLocalDate()
-            _eventStartTime.value = it.startDateTime.toLocalTime()
-            _eventEndDate.value = it.endDateTime.toLocalDate()
-            _eventEndTime.value = it.endDateTime.toLocalTime()
-            _eventLocation.value = it.location
-            _eventId.value = it.eventId
-        }
-            .onFailure { emitValidationErrorMessage("이벤트를 불러오는데 실패하였습니다.") }
-    }
-
     private fun emitValidationErrorMessage(message: String) {
         viewModelScope.launch {
-            _eventEditEvent.emit(
+            _eventRegistrationEvent.emit(
                 EventRegistrationEvent.ValidationError(message),
             )
         }
