+++ feature/management-event/src/main/res/values/strings.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="event_registration">일정 추가</string>
+    <string name="event_edit">일정 수정</string>
+    <string name="event_registration_total_page">/ 2</string>
+    <string name="event_details_title">행사 내용 작성</string>
+    <string name="event_details_content">부원들이 알아보기 쉽도록 자세하게 작성해주세요.</string>
+    <string name="event_schedule_title">"행사 일정 작성"</string>
+    <string name="event_schedule_content">정확한 날짜와 시간 그리고 장소를 입력해주세요.</string>
+    <string name="event_title">행사 이름</string>
+    <string name="event_title_hint">행사 이름을 입력하세요.</string>
+    <string name="event_content">행사 소개</string>
+    <string name="event_content_hint">행사를 이해하기 쉽도록 소개해주세요.</string>
+    <string name="event_location">장소</string>
+    <string name="event_date">날짜</string>
+    <string name="event_time">시간</string>
+    <string name="event_location_hint">어디에서 진행되는 일정인가요?</string>
+    <string name="start_date">시작 날짜</string>
+    <string name="start_time">시작 시간</string>
+    <string name="end_date">종료 날짜</string>
+    <string name="end_time">종료 시간</string>
+    <string name="date">날짜</string>
+    <string name="time">시간</string>
+    <string name="next">다음</string>
+    <string name="register_event">일정 등록</string>
+    <string name="cancel">취소</string>
+    <string name="select">선택</string>
+    <string name="backMonthArrowContentDescription">"이전 달을 보여주는 화살표 입니다."</string>
+    <string name="forwardMonthArrowContentDescription">"다음 달을 보여주는 화살표 입니다."</string>
+</resources>
