+++ feature/management/src/main/res/values/strings.xml
@@ -10,39 +10,16 @@
     <string name="detail_icon_description">자세히보기 버튼</string>
     <string name="survey">설문</string>
     <string name="event">일정</string>
-    <string name="event_registration">일정 추가</string>
-    <string name="event_edit">일정 수정</string>
-    <string name="event_registration_total_page">/ 2</string>
-    <string name="event_details_title">행사 내용 작성</string>
-    <string name="event_details_content">부원들이 알아보기 쉽도록 자세하게 작성해주세요.</string>
-    <string name="event_schedule_title">"행사 일정 작성"</string>
-    <string name="event_schedule_content">정확한 날짜와 시간 그리고 장소를 입력해주세요.</string>
-    <string name="event_title">행사 이름</string>
-    <string name="event_title_hint">행사 이름을 입력하세요.</string>
-    <string name="event_content">행사 소개</string>
-    <string name="event_content_hint">행사를 이해하기 쉽도록 소개해주세요.</string>
-    <string name="event_location">장소</string>
-    <string name="event_date">날짜</string>
-    <string name="event_time">시간</string>
-    <string name="event_location_hint">어디에서 진행되는 일정인가요?</string>
     <string name="event_date_hint">언제 진행되는 일정인가요?</string>
     <string name="event_time_hint">몇시부터 몇시까지 진행되는 일정인가요?</string>
-    <string name="register_event">일정 등록</string>
     <string name="edit_complete">수정 완료</string>
     <string name="survey_registeration">설문 등록</string>
     <string name="survey_registration_total_page">/ 4</string>
     <string name="add_survey_question">문항 추가</string>
     <string name="register_survey">설문 등록하기</string>
-    <string name="cancel">취소</string>
-    <string name="select">선택</string>
+    <string name="event_registration">일정 추가</string>
     <string name="survey_deadline_title">설문 기간 설정</string>
     <string name="survey_deadline_content">설문의 마감 기한을 설정하세요.</string>
-    <string name="start_date">시작 날짜</string>
-    <string name="start_time">시작 시간</string>
-    <string name="end_date">종료 날짜</string>
-    <string name="end_time">종료 시간</string>
-    <string name="date">날짜</string>
-    <string name="time">시간</string>
     <string name="survey_information_title">설문 제목 입력</string>
     <string name="survey_information_content">설문을 이해하기 쉽도록 제목과 소개를 입력해주세요.</string>
     <string name="survey_title">설문 제목</string>
@@ -67,7 +44,4 @@
     <string name="event_selection_content">설문을 등록할 행사를 선택하세요.</string>
     <string name="check_survey">설문 확인</string>
     <string name="name">이름</string>
-    <string name="next">다음</string>
-    <string name="backMonthArrowContentDescription">"이전 달을 보여주는 화살표 입니다."</string>
-    <string name="forwardMonthArrowContentDescription">"다음 달을 보여주는 화살표 입니다."</string>
 </resources>
