+++ settings.gradle.kts
@@ -31,5 +31,6 @@ include(":feature:profile")
 include(":feature:management")
 include(":feature:splash")
 include(":feature:management-survey")
+include(":feature:management-event")
 include(":feature:survey-check")
 include(":management-event")
